Conflicts:
	src/main/java/com/mybus/controller/LayoutController.java
	src/main/webapp/js/app.js

+++ src/main/webapp/js/services/cityManager.js
@@ -50,16 +50,14 @@ portalApp.factory('cityManager', function ($rootScope, $http, $log, $window) {
         return value.id === id;
       }));
     },
-    createCity : function (city, callback) {
-      $http.post('/api/v1/city', city)
+    createCity : function (city,callback) {
+      $http.post('/api/v1/city',city)
           .success(function (data) {
             callback(data);
-            this.fetchAllCities();
+              swal("Great", "Your City has been successfully added", "success");
           })
           .error(function (err) {
-            var errorMsg = "error adding new city info. " + (err && err.error ? err.error : '');
-            $log.error(errorMsg);
-            alert(errorMsg);
+              sweetAlert("Oops","error adding new city info.","error",+err);
           });
     },
     getCity: function (id, callback) {
@@ -72,46 +70,66 @@ portalApp.factory('cityManager', function ($rootScope, $http, $log, $window) {
             alert("error finding city. " + angular.toJson(error));
           });
     },
-    deleteCity: function(id, callback) {
-      $http.delete('/api/v1/city/' + id)
-          .success(function (data) {
-            callback(data);
-            $window.location = "#/cities";
-          })
-          .error(function (error) {
-            alert("error finding city. " + angular.toJson(error));
+      deleteCity: function(id) {
+
+          swal({   title: "Are you sure?",   text: "You will not be able to recover this City !",   type: "warning",
+                  showCancelButton: true,
+                  confirmButtonColor: "#DD6B55",
+                  confirmButtonText: "Yes, delete it!",
+                  closeOnConfirm: false }, function() {
+              $http.delete('/api/v1/city/' + id)
+                  .success(function (data) {
+                      //callback(data);
+                      sweetAlert("Great", "Your City has been successfully deleted", "success");
+                      $window.location = "#/cities";
+                  })
+                  .error(function (error) {
+                      sweetAlert("Oops...", "Error finding City data!", "error" + angular.toJson(error));
+                  });
           });
-    },
+              },
     updateCity: function(city,callback) {
       $http.put('/api/v1/city/'+city.id,city).success(function (data) {
         callback(data);
+          sweetAlert("Great","Your City has been successfully updated", "success");
         $rootScope.$broadcast('updateCityCompleteEvent');
       }).error(function (error) {
-              alert("error updating city. " + angular.toJson(error));
+          sweetAlert("Oops..", "Error updating City data!", "error" + angular.toJson(error));
           })
     },
     //----------------------------------------------------------------------
-    createBordingPoint: function (cityId,boardingPoint, callback) {
+    createBordingPoint: function (cityId,boardingPoint,callback) {
       $http.post('/api/v1/city/'+cityId+'/boardingpoint',boardingPoint).success(function (data) {
         callback(data);
-      }).error(function () {
-        alert("Error saving Bp data");
+          sweetAlert("Great","Your BoardingPoint has been successfully added", "success");
+      }).error(function (err) {
+          sweetAlert("Oops...", "Error creating Bp data!", "error"+err.message);
       });
     },
     updateBp: function(cityId,boardingPoint,callback) {
       $http.put('/api/v1/city/'+cityId+'/boardingpoint',boardingPoint).success(function (data) {
         callback(data);
+          sweetAlert("Great","Your BoardingPoint has been successfully updated", "success");
        // $rootScope.$broadcast('updateBpCompleteEvent');
       }).error(function () {
-        alert("Error updating Bp data");
+          sweetAlert("Oops...", "Error updating Bp data!", "error");
       });
     },
     deleteBp: function(cityId,BpId,callback) {
-      $http.delete('/api/v1/city/'+cityId+'/boardingpoint/'+BpId).success(function (data) {
-        callback(data);
-        //$rootScope.$broadcast('deleteBpCompleteEvent');
-      }).error(function () {
-        alert("Error deleting Bp data");
+      swal({   title: "Are you sure?",   text: "You will not be able to recover this BoardingPoint !",   type: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#DD6B55",
+          confirmButtonText: "Yes, delete it!",
+          closeOnConfirm: false }, function() {
+              $http.delete('/api/v1/city/'+cityId+'/boardingpoint/'+BpId)
+              .success(function (data) {
+                  callback(data);
+                  sweetAlert("Great", "Your BoardingPoint has been successfully deleted", "success");
+                      $rootScope.$broadcast('deleteBpCompleteEvent');
+              })
+              .error(function (error) {
+                  sweetAlert("Oops...", "Error finding City data!", "error" + angular.toJson(error));
+              });
       });
     },
     getBp: function (id,BpId, callback) {
@@ -120,7 +138,7 @@ portalApp.factory('cityManager', function ($rootScope, $http, $log, $window) {
             callback(data);
           })
           .error(function (error) {
-            alert("error finding city and Bp. " + angular.toJson(error));
+              sweetAlert("Oops...", "Error finding BoardingPoint data!", "error" + angular.toJson(error));
           });
     },
   }
